requirements.txt
fastapi
uvicorn
jinja2
sqlmodel
bcrypt
python-multipart
itsdangerous
python-dotenv
alembic
pytest
httpx
